# docs/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "이것은 Project. Stock Exchange API 문서입니다.",
        "title": "Project. Stock Exchange API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/0ghost0-dev/PJSe/blob/master/license"
        },
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/admin/symbol": {
            "get": {
                "description": "상장된 모든 심볼의 리스트를 반환합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Market - Status"
                ],
                "summary": "상장된 모든 심볼 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {API_KEY}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "성공 시 심볼 목록 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/postgresql.Symbol"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "인증 실패 시 에러 메시지 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "서버 오류 발생 시 에러 메시지 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "새로운 심볼을 상장합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin - Symbol"
                ],
                "summary": "심볼 상장",
                "parameters": [
                    {
                        "description": "상장할 심볼 정보",
                        "name": "symbol",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/postgresql.Symbol"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Bearer {API_KEY}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "성공 시 상장된 심볼 정보 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/postgresql.Symbol"
                            }
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 시 에러 메시지 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "인증 실패 시 에러 메시지 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "서버 오류 발생 시 에러 메시지 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/admin/symbol/{symbol}": {
            "get": {
                "description": "특정 심볼의 정보를 반환합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Market - Status"
                ],
                "summary": "특정 심볼 정보 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "심볼 (예: AAPL)",
                        "name": "symbol",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer {API_KEY}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "성공 시 심볼 상세 정보 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/postgresql.Symbol"
                            }
                        }
                    },
                    "401": {
                        "description": "인증 실패 시 에러 메시지 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "심볼을 찾을 수 없을 때 에러 메시지 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "서버 오류 발생 시 에러 메시지 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "심볼을 상장 폐지 상태로 변경합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin - Symbol"
                ],
                "summary": "심볼 상장 폐지",
                "parameters": [
                    {
                        "type": "string",
                        "description": "심볼 (예: AAPL)",
                        "name": "symbol",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "상장 폐지 사유",
                        "name": "reason",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Bearer {API_KEY}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "성공 시 상장 폐지 메시지 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 시 에러 메시지 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "인증 실패 시 에러 메시지 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "서버 오류 발생 시 에러 메시지 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/admin/symbol/{symbol}/now": {
            "get": {
                "description": "특정 심볼의 현재가 정보를 반환합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Market - Status"
                ],
                "summary": "특정 심볼 현재가 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "심볼 (예: AAPL)",
                        "name": "symbol",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer {API_KEY}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "성공 시 심볼 현재가 정보 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "인증 실패 시 에러 메시지 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "심볼을 찾을 수 없을 때 에러 메시지 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "서버 오류 발생 시 에러 메시지 반환",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "postgresql.Status": {
            "type": "object",
            "properties": {
                "reason": {
                    "description": "거래정지 혹은 상장폐지 사유",
                    "type": "string"
                },
                "status": {
                    "description": "\"active\", \"inactive\", \"init\", \"suspended\", \"delisted\"",
                    "type": "string"
                }
            }
        },
        "postgresql.Symbol": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "logo": {
                    "type": "string"
                },
                "market": {
                    "type": "string"
                },
                "minimum_order_quantity": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/postgresql.Status"
                },
                "symbol": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tick_size": {
                    "type": "number"
                },
                "total_stocks": {
                    "type": "integer"
                },
                "type": {
                    "description": "\"stock\", \"index\" 등",
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
    }
}